Fixed some calculations

+++ headers/hash_function.h
@@ -7,11 +7,11 @@
 class Hash_Function
 {
     private:
-        int k, w, M, d; // d is dimension of the vectors
-        unsigned long **s;
+        unsigned long k, w, M, d; // d is dimension of the vectors
+        double **s;
     public:
-        Hash_Function(int k, int w, int d);
+        Hash_Function(unsigned long k, unsigned long w, unsigned long d);
         // Calculate's hi
-        unsigned long hash(Image *image, int k);
+        unsigned long hash(Image *image, unsigned long k);
         ~Hash_Function();
 };
